In the midst of moving FishCrafter -> FishLoader
Refactoring so that all code will use LoadedFish and LoadedFishList
classes instead of Fish.  LoadedFish and LoadedFishList will
incorporate the checking that goes on in FishCrafter, and I will want
to use them everywhere to ensure that all Fish in the system are indeed
valid ones that are stored in the database, and not ones that somehow
got made up.